requirements.txt
streamlit==1.28.1
openai==1.3.7
google-generativeai==0.3.2
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
jinja2==3.1.2
markdown==3.5.1
plotly==5.17.0
pandas==2.1.3
numpy==1.25.2 